Extract some methods to helper_function namespace
